ADD: Better handling of not founding entities with OneOf

+++ src/Boilerplate.Application/Features/Users/DeleteUser/DeleteUserHandler.cs
@@ -1,12 +1,13 @@
 ﻿using Boilerplate.Application.Common;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using OneOf;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Boilerplate.Application.Features.Users.DeleteUser;
 
-public class DeleteUserHandler : IRequestHandler<DeleteUserRequest, bool>
+public class DeleteUserHandler : IRequestHandler<DeleteUserRequest, OneOf<bool, UserNotFound>>
 {
     private readonly IContext _context;
 
@@ -15,11 +16,11 @@ public class DeleteUserHandler : IRequestHandler<DeleteUserRequest, bool>
         _context = context;
     }
 
-    public async Task<bool> Handle(DeleteUserRequest request, CancellationToken cancellationToken)
+    public async Task<OneOf<bool, UserNotFound>> Handle(DeleteUserRequest request, CancellationToken cancellationToken)
     {
         var user = await _context.Users
             .FirstOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
-        // TODO: This can cause an exception. Replace this with Oneof<bool, notFound>.
+        if (user is null) return new UserNotFound();
         _context.Users.Remove(user!);
         return await _context.SaveChangesAsync(cancellationToken) > 0;
     }
